refactor(FE): #63 eslint simple-import-sort 추가 및 eslint fix 적용, UserInfo 컴포넌트 파일 삭제 (Sidebar 컴포넌트 안으로 이동)
