+++ frontend/src/config/apollo.ts
@@ -1,5 +1,6 @@
-import { ApolloClient, InMemoryCache, from, HttpLink } from "@apollo/client";
+import { ApolloClient, from, HttpLink, InMemoryCache } from "@apollo/client";
 import { onError } from "@apollo/client/link/error";
+
 import { logout } from "../util";
 
 // ref : https://www.apollographql.com/docs/react/networking/advanced-http-networking
