+++ frontend/src/layouts/AdminLayout.tsx
@@ -1,5 +1,6 @@
 import { Box } from "@chakra-ui/react";
 import { Outlet } from "react-router-dom";
+
 import Sidebar from "../components/Sidebar";
 
 const AdminLayout = () => (
